+++ jaxili/inference/npe.py
@@ -4,7 +4,11 @@ NPE.
 This module provides the Neural Posterior Estimation (NPE) class to train a neural density estimator to perform NPE.
 """
 
+import os
+import json
+import re
 import warnings
+import copy
 from typing import Any, Callable, Dict, Iterable, Optional, Union
 
 import distrax
@@ -21,11 +25,10 @@ from jaxili.loss import loss_nll_npe
 from jaxili.model import (
     ConditionalMAF,
     ConditionalRealNVP,
-    Identity,
     MixtureDensityNetwork,
     NDE_w_Standardization,
-    Standardizer,
 )
+from jaxili.compressor import Identity, Standardizer
 from jaxili.posterior import DirectPosterior
 from jaxili.train import TrainerModule
 from jaxili.utils import *
@@ -34,6 +37,7 @@ from jaxili.utils import (
     create_data_loader,
     validate_theta_x,
 )
+from jaxili.inventory.func_dict import jaxili_loss_dict, jax_nn_dict, jaxili_nn_dict
 
 default_maf_hparams = {
     "n_layers": 5,
@@ -283,7 +287,9 @@ class NPE:
         self,
         z_score_theta: bool = True,
         z_score_x: bool = True,
-        embedding_net: nn.Module = Identity(),
+        embedding_net: nn.Module = Identity,
+        embedding_hparams: dict = None,
+        **kwargs,
     ):
         """
         Build the neural network for the density estimator.
@@ -327,6 +333,8 @@ class NPE:
         if z_score_theta:
             shift = jnp.mean(self._train_dataset[:][0], axis=0)
             scale = jnp.std(self._train_dataset[:][0], axis=0)
+            min_std = kwargs.get("min_std", 1e-14)
+            scale = scale.at[scale < min_std].set(min_std)
 
         self._transformation_hparams = {"shift": shift, "scale": scale}
 
@@ -336,14 +344,32 @@ class NPE:
         if z_score_x:
             shift = jnp.mean(self._train_dataset[:][1], axis=0)
             scale = jnp.std(self._train_dataset[:][1], axis=0)
+            min_std = kwargs.get("min_std", 1e-14)
+            scale = scale.at[scale < min_std].set(min_std)
             standardizer = Standardizer(shift, scale)
         else:
             standardizer = Identity()
 
+        if embedding_net == Identity:
+            embedding_net = Identity()
+        else:
+            if embedding_hparams is None:
+                warnings.warn(
+                    "An embedding net has been specified but not its hyperparameters. Creating an embedding of the instance `Identity` instead."
+                )
+                embedding_net = Identity()
+            else:
+                embedding_net = embedding_net(**embedding_hparams)
+
         self._embedding_net = nn.Sequential([standardizer, embedding_net])
 
+        if isinstance(embedding_net, Identity):
+            n_cond = self._dim_cond
+        else:
+            n_cond = embedding_net.output_size
+
         self._model_hparams["n_in"] = self._dim_params
-        self._model_hparams["n_cond"] = self._dim_cond
+        self._model_hparams["n_cond"] = n_cond
         self._nde = self._model_class(**self._model_hparams)
 
         model = NDE_w_Standardization(
@@ -388,11 +414,13 @@ class NPE:
         except AttributeError:
             z_score_theta = kwargs.get("z_score_theta", True)
             z_score_x = kwargs.get("z_score_x", True)
-            embedding_net = kwargs.get("embedding_net", Identity())
+            embedding_net = kwargs.get("embedding_net", Identity)
+            embedding_net_hparams = kwargs.get("embedding_hparams", None)
             _ = self._build_neural_network(
                 z_score_theta=z_score_theta,
                 z_score_x=z_score_x,
                 embedding_net=embedding_net,
+                embedding_hparams=embedding_net_hparams,
             )
 
         nde_w_std_hparams = {
@@ -419,11 +447,25 @@ class NPE:
             check_val_every_epoch=check_val_every_epoch,
         )
 
-        self.trainer.config.update({"nde_hparams": self._model_hparams})
+        self.trainer.config.update({"nde_hparams": copy.deepcopy(self._model_hparams)})
+        # Check if there is an activation function to rename
+        if "activation" in self._model_hparams.keys():
+            self.trainer.config["nde_hparams"]["activation"] = self.trainer.config[
+                "nde_hparams"
+            ]["activation"].__name__
         self.trainer.config.update(
-            {"transformation_hparams": self._transformation_hparams}
+            {"transformation_hparams": copy.deepcopy(self._transformation_hparams)}
         )
-        self.trainer.init_logger(logger_params)
+        if embedding_net_hparams is not None:
+            self.trainer.config.update(
+                {"embedding_hparams": copy.deepcopy(embedding_net_hparams)}
+            )
+            # Check if there is an activation function to rename
+            if "activation" in embedding_net_hparams.keys():
+                self.trainer.config["embedding_hparams"]["activation"] = (
+                    self.trainer.config["embedding_hparams"]["activation"].__name__
+                )
+        self.trainer.write_config(self.trainer.log_dir)
 
     def train(
         self,
@@ -449,6 +491,20 @@ class NPE:
             Maximum number of epochs to train. Default is 2**31 - 1.
         check_val_every_epoch: int, optional
             Frequency at which to check the validation loss. Default is 1.
+        **kwargs : dict, optional
+            Additional keyword arguments for training customization:
+
+            - optimizer_name (str): Name of the optimizer to use (default: 'adam').
+            - gradient_clip (float): Value for gradient clipping (default: 5.0).
+            - warmup (float): Warmup proportion for learning rate scheduling (default: 0.1).
+            - weight_decay (float): Weight decay (L2 regularization) (default: 0.0).
+            - checkpoint_path (str): Directory to save training checkpoints (default: 'checkpoints/').
+            - log_dir (str or None): Directory for logging (default: None).
+            - logger_type (str): Type of logger to use (default: 'TensorBoard').
+            - seed (int): Random seed for reproducibility (default: 42).
+            - debug (bool): Whether to run in debug mode (default: False).
+            - min_delta (float): Minimum change in validation loss to qualify as improvement (default: 1e-3).
+
 
         Returns
         -------
@@ -480,6 +536,11 @@ class NPE:
                 **kwargs,
             )
         except AttributeError:
+            test_optimizer_hparams = kwargs.get("optimizer_hparams", None)
+            if test_optimizer_hparams is not None:
+                warnings.warn(
+                    "The optimizer hyperparameters specified will not be taken into account. Please refer to the documentation to modify it. Falling back to default optimizer hyperparameters."
+                )
             optimizer_hparams = {
                 "lr": learning_rate,
                 "optimizer_name": kwargs.get("optimizer_name", "adam"),
@@ -556,3 +617,151 @@ class NPE:
             )
 
         return posterior
+
+    @classmethod
+    def load_from_checkpoints(
+        cls, checkpoint: str, exmp_input: Any, embedding_net_class=Identity
+    ) -> Any:
+        """
+        Create a NPE object where the TrainerModule is loading the already existing weights for the neural network.
+
+        Parameters
+        ----------
+        nde_class: NDENetwork
+            Class used to create the neural density estimator
+        checkpoint: str
+            Folder in which the checkpoint and hyperparameter file is stored
+        exmp_input : Any
+            An input to the model with which the shapes are inferred.
+        embedding_net_class: nn.Module
+            Class used to create the embedding net. (Default: Identity)
+
+        Returns
+        -------
+        A NPE object containing a model with the pre-trained weights loaded.
+        """
+        hparams_file = os.path.join(checkpoint, "hparams.json")
+        assert os.path.isfile(hparams_file), "Could not find hparams file."
+        with open(hparams_file, "r") as f:
+            hparams = json.load(f)
+        assert (
+            hparams["model_class"] == NDE_w_Standardization.__name__
+        ), "The model has not been trained with NDE_w_Standardization. Check the checkpoint path is correct."
+        hparams.pop("model_class")
+
+        # Check that the embedding class name is correct.
+        embedding_str = hparams["model_hparams"]["embedding_net"]
+        # Find all class names in the layers list
+        class_names = re.findall(r"(\w+)\s*\(", embedding_str)
+
+        # The first entry is "Sequential", so we take the next two
+        embedding_classes = [
+            class_ for class_ in class_names[1:] if class_ != "Array"
+        ]  # Skip "Sequential"
+        assert (
+            embedding_classes[1] == embedding_net_class.__name__
+        ), "The embedding class does not match. Check that you are using the correct architecture."
+
+        # Check if the loss function is correct.
+        assert (
+            hparams["loss_fn"] in jaxili_loss_dict
+        ), "Unknown loss function. Check that the loss function you used comes from `jax.nn`."
+        hparams["loss_fn"] = jaxili_loss_dict[hparams["loss_fn"]]
+        # Create the NDE
+        # Extract the nde string
+        nde_str = hparams["model_hparams"]["nde"]
+
+        # Use regex to extract the class name
+        nde_class_match = re.match(r"(\w+)\s*\(", nde_str)
+
+        # Get the class name
+        nde_class_name = nde_class_match.group(1) if nde_class_match else None
+
+        nde_class = jaxili_nn_dict[nde_class_name]
+        nde_hparams = hparams["nde_hparams"]
+        if "activation" in nde_hparams.keys():
+            nde_hparams["activation"] = jax_nn_dict[nde_hparams["activation"]]
+
+        # Create object from the class NPE
+        inference = cls(
+            model_class=nde_class, model_hparams=nde_hparams, loss_fn=hparams["loss_fn"]
+        )
+
+        # Create the NDE
+        inference._nde = nde_class(**nde_hparams)
+
+        # Regenerate the embedding net
+        if embedding_classes[0] == "Identity":
+            standardizer = Identity()
+        elif embedding_classes[0] == "Standardizer":
+            embedding_net_str = hparams["model_hparams"]["embedding_net"]
+
+            # Regular expressions to extract mean and std arrays
+            mean_match = re.search(r"mean\s*=\s*Array\((\[.*?\])", embedding_net_str)
+            std_match = re.search(r"std\s*=\s*Array\((\[.*?\])", embedding_net_str)
+
+            # Convert extracted values into NumPy arrays
+            mean_array = (
+                np.fromstring(mean_match.group(1).strip("[]"), sep=", ")
+                if mean_match
+                else None
+            )
+            std_array = (
+                np.fromstring(std_match.group(1).strip("[]"), sep=", ")
+                if std_match
+                else None
+            )
+
+            standardizer = Standardizer(mean=mean_array, std=std_array)
+        else:
+            raise ValueError(
+                "The first class of the embedding net should be `Identity` or `Standardizer`."
+            )
+
+        if embedding_classes[1] != "Identity":
+            if "embedding_hparams" not in hparams.keys():
+                raise ValueError(
+                    "The embedding net hyperparameters can't be find. Check that you are using the correct checkpoint path."
+                )
+            if "activation" in hparams["embedding_hparams"].keys():
+                hparams["embedding_hparams"]["activation"] = jax_nn_dict[
+                    hparams["embedding_hparams"]["activation"]
+                ]
+            embedding_net = embedding_net_class(**hparams["embedding_hparams"])
+        else:
+            embedding_net = Identity()
+
+        inference._embedding_net = nn.Sequential(layers=[standardizer, embedding_net])
+
+        # Regenerate the transformation of the parameters
+        shift_str = hparams["transformation_hparams"]["shift"]
+        shift_list = [float(x) for x in shift_str.strip("[]").split()]
+
+        scale_str = hparams["transformation_hparams"]["scale"]
+        scale_list = [float(x) for x in scale_str.strip("[]").split()]
+
+        inference._transformation = distrax.ScalarAffine(
+            np.array(shift_list), np.array(scale_list)
+        )
+
+        model_hparams = {
+            "nde": inference._nde,
+            "embedding_net": inference._embedding_net,
+            "transformation": inference._transformation,
+        }
+
+        if not hparams["logger_params"]:
+            hparams["logger_params"] = dict()
+        hparams["logger_params"]["log_dir"] = checkpoint
+        hparams.pop("model_hparams")
+
+        inference.trainer = TrainerModule(
+            model_class=NDE_w_Standardization,
+            exmp_input=exmp_input,
+            model_hparams=model_hparams,
+            **hparams,
+        )
+
+        inference.trainer.load_model()
+
+        return inference