+++ jaxili/train.py
@@ -114,11 +114,7 @@ class TrainerModule:
         self.create_jitted_functions()
         self.init_model(self.exmp_input)
         # Initialize checkpointer
-        options = ocp.CheckpointManagerOptions(max_to_keep=1, create=True)
-        orbax_checkpointer = ocp.PyTreeCheckpointer()
-        self.checkpoint_manager = ocp.CheckpointManager(
-            self.log_dir, orbax_checkpointer, options
-        )
+        self.init_checkpointer()
 
     def init_logger(self, logger_params: Optional[Dict] = None):
         """
@@ -155,11 +151,20 @@ class TrainerModule:
         if not os.path.isfile(os.path.join(log_dir, "hparams.json")):
             os.makedirs(os.path.join(log_dir, "metrics/"), exist_ok=True)
             try:
-                with open(os.path.join(log_dir, "hparams.json"), "w") as f:
-                    json.dump(self.config, f, indent=4, default=handle_non_serializable)
+                self.write_config(log_dir)
             except:
                 warnings.warn("Could not save hyperparameters.", Warning)
         self.log_dir = log_dir
+
+    def write_config(self, log_dir):
+        """Write the config of the trainer in a JSON file."""
+        with open(os.path.join(log_dir, "hparams.json"), "w") as f:
+            json.dump(self.config, f, indent=4, default=handle_non_serializable)
+
+    def init_checkpointer(self):
+        """Initialize the checkpointer to save the model."""
+        options = ocp.CheckpointManagerOptions(max_to_keep=1, create=True)
+        self.checkpoint_manager = ocp.CheckpointManager(self.log_dir, options=options)
 
     def init_model(self, exmp_input: Any):
         """
@@ -205,6 +210,7 @@ class TrainerModule:
             "debug": self.debug,
             "check_val_every_epoch": self.check_val_every_epoch,
             "seed": self.seed,
+            "nde_class": self.nde_class,
         }
 
         if "activation" in self.model_hparams.keys():
@@ -415,7 +421,7 @@ class TrainerModule:
                     )
         # Test best model if possible
         if test_loader is not None:
-            # self.load_model()
+            self.load_model()
             test_metrics = self.eval_model(test_loader, log_prefix="test/")
             self.logger.log_metrics(test_metrics, step=epoch_idx)
             self.save_metrics("test", test_metrics)
@@ -607,8 +613,8 @@ class TrainerModule:
             Index of the step to save the model at, e.g. epoch.
         """
         target = {"params": self.state.params, "batch_stats": self.state.batch_stats}
-        save_args = orbax_utils.save_args_from_target(target)
-        self.checkpoint_manager.save(step, target, save_kwargs={"save_args": save_args})
+        self.checkpoint_manager.save(step, args=ocp.args.StandardSave(target))
+        self.checkpoint_manager.wait_until_finished()
 
     def load_model(self):
         """Load model and batch statistics from the logging directory."""